requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.1.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
shapely==2.0.2
pyproj==3.6.1
JinJa2==3.1.2
aiofiles==23.2.1
tables==3.9.1
h5py==3.10.1
reportlab==4.0.8
pyshp==2.3.1
scikit-image==0.26.0
slowapi==0.1.9
Pillow>=10.0.0
